refactor: move bundleTool to swingset-vat
