+++ packages/inter-protocol/test/reserve/test-reserve.js
@@ -3,15 +3,14 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { test } from '@agoric/zoe/tools/prepare-test-env-ava.js';
 
-import { E } from '@endo/eventual-send';
-import { makeIssuerKit, AmountMath } from '@agoric/ertp';
-import buildManualTimer from '@agoric/zoe/tools/manualTimer.js';
+import { AmountMath, makeIssuerKit } from '@agoric/ertp';
+import { unsafeMakeBundleCache } from '@agoric/swingset-vat/tools/bundleTool.js';
 import { eventLoopIteration } from '@agoric/zoe/tools/eventLoopIteration.js';
-
-import { setupReserveServices } from './setup.js';
-import { unsafeMakeBundleCache } from '../bundleTool.js';
+import buildManualTimer from '@agoric/zoe/tools/manualTimer.js';
+import { E } from '@endo/eventual-send';
 import { reserveInitialState, subscriptionTracker } from '../metrics.js';
 import { subscriptionKey } from '../supports.js';
+import { setupReserveServices } from './setup.js';
 
 const addLiquidPool = async (
   runPayment,
