+++ packages/wallet/contract/package.json
@@ -16,6 +16,7 @@
     "ava": "^3.12.1",
     "@agoric/ertp": "^0.14.2",
     "@agoric/inter-protocol": "^0.11.0",
+    "@agoric/swingset-vat": "^0.28.0",
     "@agoric/vats": "^0.10.0",
     "@endo/captp": "^2.0.9"
   },
